Update solidity contracts and remove executor

+++ _fuel_dev_environment/docker/deployer/fuel-v2-contracts/contracts/sidechain/types/SidechainBlockHeaderLite.sol
@@ -34,7 +34,7 @@ library SidechainBlockHeaderLiteLib {
         return
             abi.encodePacked(
                 header.prevRoot,
-                header.height,
+                (uint32)(header.height),
                 header.timestamp,
                 header.applicationHash
             );
